wrote FollowStoreImpl and reorganized class files
