+++ Greet-Dough/src/main/java/store/StorageRetrieval.java
@@ -1,4 +1,4 @@
-package utility;
+package store;
 
 import java.io.Serializable;
 import java.util.HashMap;
@@ -19,6 +19,9 @@ public abstract class StorageRetrieval<T> implements Serializable {
         this.items.put( key, newItem );
     }
 
+    // Returns:
+    //      1   if successful
+    //      0   if unsuccessful
     protected boolean delete( int ID ) {
         return ( this.items.remove(ID) != null );
     }
