+++ Greet-Dough/src/main/java/store/model/UserStoreImpl.java
@@ -1,7 +1,6 @@
-package store;
+package store.model;
 
 import model.User;
-import utility.Store;
 
 public class UserStoreImpl extends Store<User> {
